server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
requests==2.31.0
python-multipart==0.0.9
joblib==1.3.2
seaborn==0.13.2
matplotlib==3.8.2 